fix: marketplace page structure
